Moves the raw module to the top level, along with some of the API.
