+++ src/lib.rs
@@ -9,6 +9,7 @@ pub use error::{Error, Result};
 pub use value::Value;
 
 pub mod serialize;
+pub mod raw;
 
 mod blob;
 mod error;
